% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavcorrfactors.R
\name{bivCorrFacLavMod}
\alias{bivCorrFacLavMod}
\title{Generate lavaan syntax for a correlated factors solution of a bivariate
twin model. 
When running lavaan, the option group.label should be used
to indicate the value of MZ and then the DZ pair, e.g. group.label=c("MZ", "DZ").
Otherwise lavaan may mix up the twin pair correlations based on the order of entry
in the data.}
\usage{
bivCorrFacLavMod(measures, labels, model = list(A = 1, C = 1, E = 1))
}
\arguments{
\item{measures}{A named list. Measurements as names, values are list of variables to use in regressions list(meas_1=list('1, 'sex'), meas_2=list('1','sex'))}

\item{labels}{A list of labels for each measurement, eg list('phenoA', 'phenoB')}

\item{model}{A list of variance components to include, eg list(A, C=0, E=1) supresses the C component.}
}
\description{
It is assumed that measurements in the data are suffixed by "_1" and "_2"
}
