% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choleskymod.R
\name{choleskyMod}
\alias{choleskyMod}
\title{Generate a Cholesky decomposition for a multivariate twin model}
\usage{
choleskyMod(measures, suffixGroup, mod = list(A = "a", C = "c", E = "e"))
}
\arguments{
\item{measures}{A list of measurments and regressions, eg list(M1=list(1, female))}

\item{suffixGroup}{A vector of measurment suffix for MZ and DZ group.}
}
\description{
Generate a Cholesky decomposition for a multivariate twin model
}
