% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate.R
\name{correlatedFactors}
\alias{correlatedFactors}
\title{Formulate the correlated factors model}
\usage{
correlatedFactors(object)
}
\arguments{
\item{object}{The specified bivariate model.}
}
\description{
Formulate the correlated factors model
}
