% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate.R
\name{getCholeskyDefinitions}
\alias{getCholeskyDefinitions}
\title{Get parameter definitions for a bivariate cholesky model.
Contrary to the correlated factors solution, the definitions in the univariate
models cannot be used here.}
\usage{
getCholeskyDefinitions(object)
}
\description{
Get parameter definitions for a bivariate cholesky model.
Contrary to the correlated factors solution, the definitions in the univariate
models cannot be used here.
}
