% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate.R
\name{getDefinitions,Bivariate-method}
\alias{getDefinitions,Bivariate-method}
\title{Get parameter definitions for a bivariate model.}
\usage{
\S4method{getDefinitions}{Bivariate}(object)
}
\description{
Get parameter definitions for a bivariate model.
}
