% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinmodel.R
\name{getLatentFactors}
\alias{getLatentFactors}
\title{Get latent factors (ACDE)}
\usage{
getLatentFactors(object)
}
\arguments{
\item{object}{An object of class twinModel}
}
\description{
Get latent factors (ACDE)
}
