% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate.R
\name{getLoading,Bivariate,character-method}
\alias{getLoading,Bivariate,character-method}
\title{Get loadings.
This is used for the cholesky model.}
\usage{
\S4method{getLoading}{Bivariate,character}(object, fac)
}
\description{
Get loadings.
This is used for the cholesky model.
}
