% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavcholesky.R
\name{getMat}
\alias{getMat}
\title{Get one of the ACE matrixes from a fitted model in lavaan.}
\usage{
getMat(ft, measures, factor = "A")
}
\arguments{
\item{ft}{The fitted model}

\item{measures}{The names of the measurments}

\item{factor}{Which component to include.}
}
\description{
Get one of the ACE matrixes from a fitted model in lavaan.
}
