% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{getParameterLabels}
\alias{getParameterLabels}
\title{Get parameter labels}
\usage{
getParameterLabels(object)
}
\arguments{
\item{object}{An object of class twinModel}
}
\description{
Get parameter labels
}
