% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavcholesky.R
\name{getVarExp}
\alias{getVarExp}
\title{Get the variance explained}
\usage{
getVarExp(ft, measures, factor = "A")
}
\arguments{
\item{ft}{The fitted lavaan model.}

\item{measures}{A list of the measurement names.}

\item{factor}{The factor to extract.}
}
\description{
Get the variance explained
}
