% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinmodel.R
\name{regress}
\alias{regress}
\title{Add regression to the twin model}
\usage{
regress(object, reg)
}
\arguments{
\item{reg}{A regression formula for a measurement without suffix for measurements in twin and co-twin.}
}
\description{
Add regression to the twin model
}
