% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{setLatentFactors}
\alias{setLatentFactors}
\title{Set latent factors (ACDE)}
\usage{
setLatentFactors(object, ...)
}
\arguments{
\item{...}{Each factor specified as factor=label, eg A='a'}
}
\description{
Set latent factors (ACDE)
}
