% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinmodel.R
\name{setSuffix}
\alias{setSuffix}
\title{Set the suffix for measurments in twin and co-twin}
\usage{
setSuffix(object, t1, t2)
}
\arguments{
\item{t1}{Measurement suffix for twin 1}

\item{t2}{Measurement suffix for twin 2}
}
\description{
Set the suffix for measurments in twin and co-twin
}
