% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavcholesky.R
\name{trivCholLavMod}
\alias{trivCholLavMod}
\title{Trivariate cholesky model
Pass measurements in an indexed list, variable names as indexes and regressions as values, eg:
list(var1="c(int,int)*1 + ...", var2="c(int,int)*1 + ...", var3="c(int,int)*1 + ...")}
\usage{
trivCholLavMod(measures)
}
\arguments{
\item{measures}{The list of measurements and regressions.}
}
\description{
Trivariate cholesky model
Pass measurements in an indexed list, variable names as indexes and regressions as values, eg:
list(var1="c(int,int)*1 + ...", var2="c(int,int)*1 + ...", var3="c(int,int)*1 + ...")
}
