% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinmodel.R
\docType{class}
\name{twinModel-class}
\alias{twinModel-class}
\title{The twinModel class}
\description{
The twinModel class
}
\section{Slots}{

\describe{
\item{\code{factors}}{A, C, D, E factors in a twin model.}

\item{\code{regressions}}{Regressions to perform on measurements}

\item{\code{suffix}}{Suffix for the measurement in twin and co twin}
}}

