% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\docType{class}
\name{Univariate-class}
\alias{Univariate-class}
\title{The univariate class}
\description{
The univariate class
}
\section{Slots}{

\describe{
\item{\code{factors}}{A, C, D, E factors in a twin model.}

\item{\code{regressions}}{Regressions to perform on measurements}

\item{\code{suffix}}{Suffix for the measurement in twin and co twin}

\item{\code{latentSuffix}}{Suffix for latent variables.}
}}

