% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univtwinmod.R
\name{univTwinMod}
\alias{univTwinMod}
\title{Generate syntax for a univariate twin model}
\usage{
univTwinMod(
  measT1,
  measT2,
  model = list(A = "a", C = "c", E = "e"),
  append = "",
  varLabels = list(A = "add", C = "com", D = "dom", E = "uni"),
  independentT1 = list(int = "1"),
  independentT2 = list(int = "1")
)
}
\arguments{
\item{measT1}{Variable name of measurement in twin one.}

\item{measT2}{Variable name of measurement in twin two.}

\item{model}{A list of variance components (A,C,D,E) and labels, eg list(A='a',C='c',E='e').}

\item{append}{Any additional syntax.}

\item{varLabels}{Currently not used.}
}
\description{
Generate syntax for a univariate twin model
}
