% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivtwinmod.R
\name{bivTwinMod}
\alias{bivTwinMod}
\title{Generate syntax for a bivariate twin model}
\usage{
bivTwinMod(
  measT1,
  measT2,
  regressions = list(list(int = 1, fem = "female"), list(int = 1, fem = "female")),
  model = list(A = "a", C = "c", E = "e")
)
}
\arguments{
\item{measT1}{A list of variable name and a designated label eg label(varName1="label1") for measure 1.}

\item{measT2}{A list of variable name and a designated label eg label(varName1="label2") for measure 2.}

\item{regressions}{A list of lists for the regressions for measT1 and measT1.}

\item{model}{A list of variance components (A,C,D,E) and labels, eg list(A='a',C='c',E='e').}
}
\description{
Generate syntax for a bivariate twin model
}
